+++ dynamodb/src/test/java/docs/javadsl/RetryTest.java
@@ -12,9 +12,9 @@ import akka.stream.ActorMaterializer;
 import akka.stream.Materializer;
 import akka.stream.alpakka.dynamodb.AwsOp;
 import akka.stream.alpakka.dynamodb.ItemSpecOps;
+import akka.stream.alpakka.dynamodb.impl.javadsl.RetryFlow;
 import akka.stream.alpakka.dynamodb.javadsl.DynamoDb;
 import akka.stream.javadsl.Flow;
-import akka.stream.javadsl.RetryFlow;
 import akka.stream.javadsl.Source;
 import akka.stream.testkit.javadsl.StreamTestKit;
 import com.amazonaws.services.dynamodbv2.model.BatchGetItemResult;
